fixed simple api

+++ events/views_api.py
@@ -9,12 +9,6 @@ from .serializers import EventSerializer
 class EventViewSet(viewsets.ViewSet):
 
     def list(self, request):
-        """
-        GET
-
-        Example:
-            http://127.0.0.1:8000/api/events/list/?title=Субб&category=3
-        """
         queryset = Event.objects_event_qs.event_qs()
         filter_category = request.query_params.get('category')
         filter_features = request.GET.getlist('features')
@@ -46,38 +40,19 @@ class EventViewSet(viewsets.ViewSet):
         return Response(serializer.data)
 
     def retrieve(self, request, pk=None):
-        """
-        GET
-
-        Example:
-            http://127.0.0.1:8000/api/events/detail/1/
-        """
         queryset = Event.objects_event_qs.event_qs()
         event = get_object_or_404(queryset, pk=pk)
         serializer = EventSerializer(event)
         return Response(serializer.data)
 
     def destroy(self, request, pk=None):
-        """
-        DELETE
-
-        Example:
-            http://127.0.0.1:8000/api/events/event_delete/37/
-        """
         event = Event.objects_event_qs.event_qs().filter(id=pk)
+        if not event:
+            return Response(status=status.HTTP_404_NOT_FOUND, data="Страница не найдена")
         event.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(status=status.HTTP_204_NO_CONTENT, data=f'Событие удалено')
 
     def partial_update(self, request, pk=None):
-        """
-        PATCH
-
-        EXAMPLE:
-            http://127.0.0.1:8000/api/events/event_update/37/
-
-        {"title": "TEST_TITLE_2",
-        "date_start": "2034-04-23T18:25:43.511Z"}
-        """
         event = Event.objects_event_qs.event_qs().get(id=pk)
         serializer = EventSerializer(event, data=request.data,
                                      partial=True)
@@ -87,18 +62,6 @@ class EventViewSet(viewsets.ViewSet):
         return Response(status=status.HTTP_400_BAD_REQUEST)
 
     def create(self, request):
-        """
-        POST
-
-        EXAMPLE:
-            http://127.0.0.1:8000/api/events/event_create/
-
-        {"title": "TEST_TITLE",
-        "date_start": "2032-04-23T18:25:43.511Z",
-        "category": 1,
-        "participants_number": 12,
-        "features": [{"id": 1}, {"id": 2}]}
-        """
         try:
             new_event = Event.objects.create(title=request.data['title'],
                                              description=request.data['description'] if request.data.get(
@@ -112,6 +75,10 @@ class EventViewSet(viewsets.ViewSet):
             if request.data.get('features'):
                 new_event.features.set(Feature.objects.filter(id__in=[i['id'] for i in request.data['features']]))
             new_event.save()
-            return Response(status=status.HTTP_201_CREATED)
-        except Exception:
-            return Response(status=status.HTTP_400_BAD_REQUEST)
+            return Response(status=status.HTTP_201_CREATED, data='Событие создано')
+        except Exception as e:
+            return Response(status=status.HTTP_400_BAD_REQUEST, data='Поля title, date_start, '
+                                                                     'participants_number, category '
+                                                                     'являются обязательными. '
+                                                                     f'При обработке запроса возникла ошибка: '
+                                                                     f'{str(e)}')
